SlurmHyperopt

Documentation for SlurmHyperopt.

DataFrames.DataFrame

SlurmHyperopt.SlurmHyperoptimizer

SlurmHyperopt.SlurmParams

Base.getindex

SlurmHyperopt.generate_slurm_file

SlurmHyperopt.merge_results!

SlurmHyperopt.save_result

DataFrames.DataFrame — Method

DataFrames(sho::SlurmHyperoptimizer)

Returns the results and parameters as a DataFrame

SlurmHyperopt.SlurmHyperoptimizer — Type

SlurmHyperoptimizer(ho::Hyperoptimizer; slurm_params::SlurmParams=nothing)

Creates a Slurm Job Array submission file and a Hyperoptimizer instance for your Hyperoptimizer needs. Takes the same arguments as Hyperoptimizer from Hyperopt + the keyword argument slurm_params.

SlurmHyperopt.SlurmParams — Type

struct SlurmParams

Holds all user defined parameters for the Slurm script. The constructor works with keyword arguments. If a field is nothing, it is omitted in the script. Hyphens "-" are replaced by underscores "_" compared to regular Slurm names. Currently:

* `qos`
* `job_name`
* `account`
* `output = nothing` if output and error are nothing, they are automatically generated based on the job name and the job array task index
* `error = nothing`
* `partition = nothing`
* `gres = nothing`
* `nodes = nothing`
* `ntasks_per_node = nothing`
* `cpus-per-task = nothing`
* `workdir = nothing`
* `mail_type = nothing`
* `mail_user = nothing`
* `parallel_jobs = nothing`: how many jobs can run at the same time?
* `file_path = ""`: path to Slurm script file to be saved
* `extra_calls = ""`: string with extra calls, e.g. loading modules etc
* `julia_call = ""`: julia call

Base.getindex — Method

get_index(ho::SlurmHyperoptimizer, i::Integer)

Very hacky way how to index a Hyperoptimizer object, that respect its history, in case the individual samples are not independend (as for Hyperband)

SlurmHyperopt.generate_slurm_file — Method

generate_slurm_file(p::SlurmParams, N_jobs::Integer)

Generates the Slurm script file based on a SlurmParams instance.

SlurmHyperopt.merge_results! — Method

merge_results!(sho::SlurmHyperoptimizer)

Merges the results and store them in the sho.results field. Also deletes the temporal files.

SlurmHyperopt.save_result — Method

save_result(sho::SlurmHyperoptimizer, res::HyperoptResults, i::Integer)

(Temporally) saves the results res of job i. Results are later merged with merge_results!.